refactor: extract prepareOutDir as a plugin
